Refactoring twiddle coefficients calculations

+++ CChirpZ.cpp
@@ -87,7 +87,7 @@ CChirpZ::CChirpZ(int Points) {
 // modulate X[n] and then zero padding to M, which is radix 2 and larger than or equal to 2N-1
 // Y[n] = X[n] * C[n], then Y[n] is written back to X[n].  
 
-void CChirpZ::ConvertFloat(double MathPiWithSign) {
+void CChirpZ::ConvertFloat(long double MathPiWithSign) {
     
     // assume input data have been stored in X from 0 to mPoints-1.
     // pad the rest with zeros
@@ -99,14 +99,14 @@ void CChirpZ::ConvertFloat(double MathPiWithSign) {
     long long int n  = 0;
     long long int n2 = 0;
     long long int mPoints2 = 2 * mPoints;
-    double Index = 0.0;
+    long double Index = 0.0L;
     
     for(int i = 0; i < mPoints; i++) {
         //n2 = (long long int)i;
         //n2 = n2 * n2;
         
         n2 = n % mPoints2;
-        Index = (double)n2 / (double)mPoints * MathPiWithSign;
+        Index = (long double)n2 / (long double)mPoints * MathPiWithSign;
         n += (2*i+1);
         C[re(i)] = 0.0f;
         C[im(i)] = (float)Index;
@@ -165,6 +165,7 @@ CChirpZ::CChirpZ(const CChirpZ& orig) {
 }
 
 CChirpZ::~CChirpZ() {
+    
     free(C);
     free(X);
     free(Cs);
