+++ CChirpZ.h
@@ -11,17 +11,20 @@
 class CChirpZ {
     
 public:
+    
     CChirpZ(int Points);
     virtual ~CChirpZ();
     
     float * GetInputFloat(void);
     float * GetOutputFloat(void);    
-    void ConvertFloat(double MathPiWithSign);  // M_PI = forward, -M_PI = backward        
+    void ConvertFloat(long double MathPiWithSign);  // M_PIl = forward, -M_PIl = backward        
     
 private:
+    
     CChirpZ(const CChirpZ& orig);
     
 protected:
+    
     int mPoints;
     int mPointsChirpZ;  // radix 2 and greater than or equal to 2 * mPoints - 1
